refactor: web layer 관련된 Controller, Dto 관련 클래스 파일 패키지 정리
- 각 도메인별로 안에서 각각 Controller, dto 패키지가 분리되어 있던 구조에서 변경
  클라이언트 통신과 가까운 controller, dto 패키지를 api 패키지 하위로 이동
